Fresh makeup for Leggero-v2 style
This elevates the Leggero-v2 style to a more modern design.
In most situations in life, first impressions count. The FlatPress Leggero-v2 style is what the user sees first after performing the setup.
Potential new users may not immediately find their way to the wiki to see the variety of themes there - and hastily delete FlatPress again. Therefore I made a few small changes to the style:

- Adjusted the alignment of the calendar widget and the search widget.
- A single PhotoSwipe image or a whole gallery is now centered in the responsive design. #150
- BBcode videos are no longer chopped off in responsive design, but adjusted to the width and center aligned.
- The overall appearance is now not so angular/edgy.
- A BBcode embedded Facebook video now has localized languages with language tag. #252
  (https://developers.facebook.com/docs/javascript/internationalization).
- BBcode: Removed reference to an active lightbox plugin, as PhotoSwipe can be used as an alternative.
- Fixed an error in the HTML code of PhotoSwipe-Plugin. #253
- Lucida Console [code] ... [/code]is now correct as a font in the CSS file.
- Admin area now has Leggero-v2 style background instead of white background.

With best regards
Frank