PhotoSwipe 4.1.3 included
This upgrades PhotoSwipe to the January 8, 2019 release level and is the last version before rewriting to ES6. Fixed the following:

Fixes an issue with devices that have multiple input types and PhotoSwipe does not wipe.
Fixes an issue where the layout does not update after an alignment change in some cases.
Previous FlatPress-specific fixes have been considered and incorporated.
Details can be found here:
https://github.com/dimsemenov/PhotoSwipe/releases

With best regards
Frank