Update header.tpl
Logic changed to: if feed plugin is deactivated, the feed buttons are displayed in the header. It is unlikely that the FlatPress user wants the buttons in the head and in the menu bar to be displayed at the same time. If the FlaPress user uses a different theme, he can still activate the plugin/widget.