Fixes #517 |  "Array to string conversion" warning
Hi @azett,

I can also reproduce it. Once with my test system without cache mechanisms, but also on Pi, with the production system. However, I firmly believe that it is not a side effect of #499. When the configuration menu page is reloaded (F5), only the currently set lang_load language is loaded.

The language arrays for the configuration menu were only assigned to the template after reloading the page until 1.3.1. For users with OPcache and/or APCu, nothing happened after clicking on “Save change”. Saved changes were only loaded after reloading the configuration page. Therefore the 18 year old problem was never noticed.
Now onsave() with setup() immediately pushes the changed keys into the template. And now we have two keys in lang_load - 0 and 1. After reloading the page, everything is fine again.

One possible cause may be that we are using two different Smarty instances ($smarty and $_FP_SMARTY). I hope that this will not be a problem when we switch to Smarty 5.

I have now fiddled with an ugly solution in the plugin_lastcomments_rsshead function. Wasn't the first, won't be the last :-D.